Defined Plugin interface
Defined an abstract persistance factory interface which creates GameDAOs and UserDAOs. The UserDAO can add a user to persistent storage and retrieve a list of all users in storage. The GameDAO does the same but can also set the number of checkpoints to cache between updating the Game and defines a method for applying a command to a game.